+++ src/draw/ft_radian_domain.c
@@ -6,14 +6,14 @@
 /*   By: ryoshio- <ryoshio-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 10:39:43 by ryoshio-          #+#    #+#             */
-/*   Updated: 2023/03/25 12:45:25 by ryoshio-         ###   ########.fr       */
+/*   Updated: 2023/03/29 00:49:19 by ryoshio-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 
 #include "../includes/cub3d.h"
 
-float ft_radian_domain(float angle)
+double ft_radian_domain(double angle)
 {
 	if (angle >= 2 * PI)
 		angle -= 2 * PI;
